pyproject.toml
[project]
name = "Slide-Generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "ipython>=8.35.0",
    "langchain>=0.3.22",
    "langchain-openai>=0.3.12",
    "langgraph-supervisor>=0.0.16",
    "pillow>=11.1.0",
    "pydantic>=2.11.1",
]
